Migrating devLUA functions to xxParameters.cpp which is part of the endpoint
