+++ src/lib/tx-crsf/devTXLUA.h
@@ -3,5 +3,3 @@
 #include "device.h"
 
 extern device_t TXLUA_device;
-
-void luadevUpdateFolderNames();
